backend/requirements.txt
fastapi
uvicorn
pymongo
graphene
python-jose
pydantic
motor
ariadne
starlette
mysql-connector-python
